+++ src/components/Button/Button.stories.tsx
@@ -2,7 +2,7 @@
 
 import { Meta, Story } from '@storybook/react';
 
-import Button, { ButtonProps } from '..';
+import Button, { ButtonProps } from '.';
 
 export default {
   title: 'Button',
