+++ src/components/CardCompany/style.ts
