+++ src/pages/CreditCardRegister/CreditCardPasswordInput/style.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const PasswordBox = styled.div`
+  width: 48px;
+  height: 48px;
+  margin-right: 10px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 1px solid #ecebf1;
+  border-radius: 7px;
+`;
+
+export default PasswordBox;
