+++ src/pages/CreditCardRegister/InputLabel/index.tsx
@@ -1,0 +1,11 @@
+import * as S from './style';
+
+type InputLabelType = {
+  label: string;
+};
+
+function InputLabel({ label }: InputLabelType) {
+  return <S.LabelWrapper>{label}</S.LabelWrapper>;
+}
+
+export default InputLabel;
