+++ src/pages/CreditCardRegister/InputLabel/style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const LabelWrapper = styled.div`
+  font-size: 12px;
+  color: #525252;
+  margin-bottom: 5px;
+`;
+
+export default {};
