+++ src/pages/CreditCardRegister/InputLayout/style.ts
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+export const InputLayout = styled.div`
+  position: relative;
+`;
+
+export const ErrorMessage = styled.div`
+  font-size: 12px;
+  position: absolute;
+  bottom: -16px;
+  font-weight: 500;
+  color: rgba(255, 50, 50, 0.8);
+`;
